data_generator/requirements.txt
kafka-python==2.0.2
python-dateutil==2.8.2
typing-extensions==4.5.0 